pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backlink-broadcast"
version = "1.1.0"
description = "AI-powered autonomous radio station using swarm intelligence"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Andon Labs", email = "dev@andonlabs.com"},
    {name = "SMTP.eth Team"},
    {name = "nft2.me"},
    {name = "fuzzywigg.ai Development Team"}
]
keywords = [
    "ai",
    "radio",
    "swarm-intelligence",
    "llm",
    "autonomous-agents",
    "broadcast"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "google-api-python-client>=2.88.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "google-genai>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "google-cloud-firestore>=2.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.0",
]
twitter = [
    "tweepy>=4.14.0",
]
browser = [
    "browser-use>=0.1.0",
]
all = [
    "backlink-broadcast[dev,twitter,browser]",
]

[project.urls]
Homepage = "https://github.com/fuzzywigg/Backlink"
Documentation = "https://github.com/fuzzywigg/Backlink#readme"
Repository = "https://github.com/fuzzywigg/Backlink.git"
Issues = "https://github.com/fuzzywigg/Backlink/issues"

[project.scripts]
backlink = "hive.queen.orchestrator:main"
backlink-cache = "hive.utils.cache_manager:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hive*", "constitutional_llm*"]
exclude = ["tests*", "test_*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.md"]

# ─────────────────────────────────────────────────────────────────────────────
# Tool Configuration
# ─────────────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests", "hive/tests", "constitutional-llm/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
    "unit: marks unit tests",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["hive"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
fail_under = 60

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict parameter
    "C901",   # too complex
    "ARG002", # Unused method argument
    "ARG004", # Unused static method argument
    "ARG001", # Unused function argument
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG001", "S101"]

[tool.ruff.lint.isort]
known-first-party = ["hive", "constitutional_llm"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["hive", "constitutional_llm"]
skip_gitignore = true
